Move cloudformation to common json parser
