+++ checkov/cloudformation/checks/resource/aws/RedShiftSSL.py
@@ -1,5 +1,5 @@
 from checkov.cloudformation.checks.resource.base_resource_check import BaseResourceCheck
-from checkov.cloudformation.parser.node import dict_node
+from checkov.common.parsers.node import dict_node
 from checkov.common.models.enums import CheckCategories, CheckResult
