+++ checkov/cloudformation/checks/resource/base_resource_value_check.py
@@ -5,7 +5,7 @@ from typing import List, Any, Dict
 from checkov.cloudformation.checks.resource.base_resource_check import BaseResourceCheck
 from checkov.cloudformation.context_parser import ContextParser
 from checkov.cloudformation.parser import dict_node
-from checkov.cloudformation.parser.node import str_node
+from checkov.common.parsers.node import str_node
 from checkov.common.models.consts import ANY_VALUE
 from checkov.common.models.enums import CheckResult, CheckCategories
 from checkov.common.util.type_forcers import force_list
