+++ checkov/cloudformation/checks/utils/iam_cloudformation_document_to_policy_converter.py
@@ -1,6 +1,6 @@
 import copy
 
-from checkov.cloudformation.parser.node import dict_node
+from checkov.common.parsers.node import dict_node
 
 
 def convert_cloudformation_conf_to_iam_policy(conf: dict_node) -> dict_node:
